refactor: move line chart components to folder
