+++ app/vmui/packages/vmui/src/components/Chart/Line/ChartTooltip/style.scss
@@ -78,5 +78,6 @@ $chart-tooltip-y: -1 * ($padding-small + $chart-tooltip-half-icon);
     display: grid;
     grid-gap: 4px;
     word-break: break-all;
+    white-space: pre-wrap;
   }
 }
