+++ app/vmui/packages/vmui/src/components/Chart/Line/Legend/Legend.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useMemo } from "preact/compat";
-import { LegendItemType } from "../../../utils/uplot/types";
+import { LegendItemType } from "../../../../utils/uplot/types";
 import LegendItem from "./LegendItem/LegendItem";
-import Accordion from "../../Main/Accordion/Accordion";
+import Accordion from "../../../Main/Accordion/Accordion";
 import "./style.scss";
 
 interface LegendProps {
