+++ app/vmui/packages/vmui/src/components/Chart/Line/Legend/LegendItem/LegendItem.tsx
@@ -1,19 +1,23 @@
 import React, { FC, useState, useMemo } from "preact/compat";
 import { MouseEvent } from "react";
-import { LegendItemType } from "../../../../utils/uplot/types";
+import { LegendItemType } from "../../../../../utils/uplot/types";
 import "./style.scss";
 import classNames from "classnames";
-import Tooltip from "../../../Main/Tooltip/Tooltip";
+import Tooltip from "../../../../Main/Tooltip/Tooltip";
 import { getFreeFields } from "./helpers";
 
 interface LegendItemProps {
   legend: LegendItemType;
-  onChange: (item: LegendItemType, metaKey: boolean) => void;
+  onChange?: (item: LegendItemType, metaKey: boolean) => void;
+  isHeatmap?: boolean;
 }
 
-const LegendItem: FC<LegendItemProps> = ({ legend, onChange }) => {
+const LegendItem: FC<LegendItemProps> = ({ legend, onChange, isHeatmap }) => {
   const [copiedValue, setCopiedValue] = useState("");
-  const freeFormFields = useMemo(() => getFreeFields(legend), [legend]);
+  const freeFormFields = useMemo(() => {
+    const result = getFreeFields(legend);
+    return isHeatmap ? result.filter(f => f.key !== "vmrange") : result;
+  }, [legend, isHeatmap]);
   const calculations = legend.calculations;
   const showCalculations = Object.values(calculations).some(v => v);
 
@@ -24,7 +28,7 @@ const LegendItem: FC<LegendItemProps> = ({ legend, onChange }) => {
   };
 
   const createHandlerClick = (legend: LegendItemType) => (e: MouseEvent<HTMLDivElement>) => {
-    onChange(legend, e.ctrlKey || e.metaKey);
+    onChange && onChange(legend, e.ctrlKey || e.metaKey);
   };
 
   const createHandlerCopy = (freeField: string, id: string) => (e: MouseEvent<HTMLDivElement>) => {
@@ -37,18 +41,21 @@ const LegendItem: FC<LegendItemProps> = ({ legend, onChange }) => {
       className={classNames({
         "vm-legend-item": true,
         "vm-legend-row": true,
-        "vm-legend-item_hide": !legend.checked,
+        "vm-legend-item_hide": !legend.checked && !isHeatmap,
+        "vm-legend-item_static": isHeatmap,
       })}
       onClick={createHandlerClick(legend)}
     >
-      <div
-        className="vm-legend-item__marker"
-        style={{ backgroundColor: legend.color }}
-      />
+      {!isHeatmap && (
+        <div
+          className="vm-legend-item__marker"
+          style={{ backgroundColor: legend.color }}
+        />
+      )}
       <div className="vm-legend-item-info">
         <span className="vm-legend-item-info__label">
           {legend.freeFormFields["__name__"]}
-          &#123;
+          {!!freeFormFields.length && <>&#123;</>}
           {freeFormFields.map((f, i) => (
             <Tooltip
               key={f.id}
@@ -66,10 +73,10 @@ const LegendItem: FC<LegendItemProps> = ({ legend, onChange }) => {
               </span>
             </Tooltip>
           ))}
-          &#125;
+          {!!freeFormFields.length && <>&#125;</>}
         </span>
       </div>
-      {showCalculations && (
+      {!isHeatmap && showCalculations && (
         <div className="vm-legend-item-values">
           median:{calculations.median}, min:{calculations.min}, max:{calculations.max}, last:{calculations.last}
         </div>
