+++ app/vmui/packages/vmui/src/components/Chart/Line/Legend/style.scss
